+++ src/modules/MainPage/components/CountryCard/CountryCard.tsx
