+++ src/modules/MainPage/components/CountryCard/index.ts
